test: rewrite test for refactored code
