+++ pyopenagi/tools/imdb/top_movie.py
@@ -1,4 +1,4 @@
-from ...base import BaseRapidAPITool
+from ..base import BaseRapidAPITool
 
 from typing import Any, Dict, List, Optional
