+++ tests/test_agent_creation.py
@@ -4,13 +4,13 @@ import pytest
 
 def test_agent_creation():
     agent_process = AgentProcess(
-        agent_name="Narrative Agent",
+        agent_name="example/academic_agent",
         query=Query(
             messages = [
-                {"role": "user", "content": "Craft a tale about a valiant warrior on a quest to uncover priceless treasures hidden within a mystical island."}
+                {"role": "user", "content": "Summarize researches of quantum computing in recent five years."}
             ]
         )
     )
     # Use plain assert statements for testing conditions
-    assert agent_process.agent_name == "Narrative Agent", "Agent name does not match"
+    assert agent_process.agent_name == "example/academic_agent", "Agent name does not match"
     # Add more assertions here as necessary to validate the properties of the agent_process object
