+++ tests/test_tools/test_currency_converter.py
@@ -1,7 +1,7 @@
 import os
 import pytest
 
-from pyopenagi.tools.online.currency_converter import CurrencyConverterAPI
+from pyopenagi.tools.currency_converter.currency_converter import CurrencyConverter
 from dotenv import load_dotenv, find_dotenv
 
 @pytest.fixture(scope="module")
@@ -9,7 +9,7 @@ def test_rapid_api_key():
     load_dotenv(find_dotenv())
     if "RAPID_API_KEY" not in os.environ or not os.environ["RAPID_API_KEY"]:
         with pytest.raises(ValueError):
-            currency_converter_api = CurrencyConverterAPI()
+            currency_converter = CurrencyConverter()
         pytest.skip("Rapid api key is not set.")
     else:
         return True
@@ -17,7 +17,7 @@ def test_rapid_api_key():
 @pytest.mark.usefixtures("test_rapid_api_key")
 def test_currency_converter_api():
     load_dotenv(find_dotenv())
-    currency_converter_api = CurrencyConverterAPI()
+    currency_converter_api = CurrencyConverter()
     params = {
         "from": "USD",
         "to": "EUR",
